split playwright tests in ci
